Updated folder structure: Khin-Hpone/Project-01
