--- Project_01/data/HDBRPIMonthly.csv
@@ -1,400 +1,0 @@
-month,index
-1990-01,24.3
-1990-02,24.3
-1990-03,24.3
-1990-04,24.4
-1990-05,24.4
-1990-06,24.4
-1990-07,25
-1990-08,25
-1990-09,25
-1990-10,24.7
-1990-11,24.7
-1990-12,24.7
-1991-01,24.9
-1991-02,24.9
-1991-03,24.9
-1991-04,25.5
-1991-05,25.5
-1991-06,25.5
-1991-07,25.2
-1991-08,25.2
-1991-09,25.2
-1991-10,25.1
-1991-11,25.1
-1991-12,25.1
-1992-01,25.8
-1992-02,25.8
-1992-03,25.8
-1992-04,27.4
-1992-05,27.4
-1992-06,27.4
-1992-07,28
-1992-08,28
-1992-09,28
-1992-10,28.6
-1992-11,28.6
-1992-12,28.6
-1993-01,30.2
-1993-02,30.2
-1993-03,30.2
-1993-04,39.6
-1993-05,39.6
-1993-06,39.6
-1993-07,47.7
-1993-08,47.7
-1993-09,47.7
-1993-10,49
-1993-11,49
-1993-12,49
-1994-01,50.4
-1994-02,50.4
-1994-03,50.4
-1994-04,51.7
-1994-05,51.7
-1994-06,51.7
-1994-07,54.6
-1994-08,54.6
-1994-09,54.6
-1994-10,54.8
-1994-11,54.8
-1994-12,54.8
-1995-01,57.3
-1995-02,57.3
-1995-03,57.3
-1995-04,63.9
-1995-05,63.9
-1995-06,63.9
-1995-07,67.7
-1995-08,67.7
-1995-09,67.7
-1995-10,73.7
-1995-11,73.7
-1995-12,73.7
-1996-01,80.5
-1996-02,80.5
-1996-03,80.5
-1996-04,90.9
-1996-05,90.9
-1996-06,90.9
-1996-07,96.4
-1996-08,96.4
-1996-09,96.4
-1996-10,99
-1996-11,99
-1996-12,99
-1997-01,98.6
-1997-02,98.6
-1997-03,98.6
-1997-04,97.5
-1997-05,97.5
-1997-06,97.5
-1997-07,93.6
-1997-08,93.6
-1997-09,93.6
-1997-10,88
-1997-11,88
-1997-12,88
-1998-01,81.7
-1998-02,81.7
-1998-03,81.7
-1998-04,78.3
-1998-05,78.3
-1998-06,78.3
-1998-07,74.9
-1998-08,74.9
-1998-09,74.9
-1998-10,72.3
-1998-11,72.3
-1998-12,72.3
-1999-01,71.2
-1999-02,71.2
-1999-03,71.2
-1999-04,72.2
-1999-05,72.2
-1999-06,72.2
-1999-07,78.1
-1999-08,78.1
-1999-09,78.1
-1999-10,79.8
-1999-11,79.8
-1999-12,79.8
-2000-01,80.3
-2000-02,80.3
-2000-03,80.3
-2000-04,79.3
-2000-05,79.3
-2000-06,79.3
-2000-07,77.6
-2000-08,77.6
-2000-09,77.6
-2000-10,75.8
-2000-11,75.8
-2000-12,75.8
-2001-01,73.2
-2001-02,73.2
-2001-03,73.2
-2001-04,72.1
-2001-05,72.1
-2001-06,72.1
-2001-07,70.6
-2001-08,70.6
-2001-09,70.6
-2001-10,69.6
-2001-11,69.6
-2001-12,69.6
-2002-01,69.1
-2002-02,69.1
-2002-03,69.1
-2002-04,69.2
-2002-05,69.2
-2002-06,69.2
-2002-07,69.9
-2002-08,69.9
-2002-09,69.9
-2002-10,69.9
-2002-11,69.9
-2002-12,69.9
-2003-01,71
-2003-02,71
-2003-03,71
-2003-04,72.5
-2003-05,72.5
-2003-06,72.5
-2003-07,74.3
-2003-08,74.3
-2003-09,74.3
-2003-10,75.1
-2003-11,75.1
-2003-12,75.1
-2004-01,75.3
-2004-02,75.3
-2004-03,75.3
-2004-04,76.2
-2004-05,76.2
-2004-06,76.2
-2004-07,76.3
-2004-08,76.3
-2004-09,76.3
-2004-10,77.1
-2004-11,77.1
-2004-12,77.1
-2005-01,77.2
-2005-02,77.2
-2005-03,77.2
-2005-04,73.5
-2005-05,73.5
-2005-06,73.5
-2005-07,73.2
-2005-08,73.2
-2005-09,73.2
-2005-10,73.5
-2005-11,73.5
-2005-12,73.5
-2006-01,73.6
-2006-02,73.6
-2006-03,73.6
-2006-04,74.3
-2006-05,74.3
-2006-06,74.3
-2006-07,74.2
-2006-08,74.2
-2006-09,74.2
-2006-10,74.9
-2006-11,74.9
-2006-12,74.9
-2007-01,75.8
-2007-02,75.8
-2007-03,75.8
-2007-04,78.1
-2007-05,78.1
-2007-06,78.1
-2007-07,83.2
-2007-08,83.2
-2007-09,83.2
-2007-10,88
-2007-11,88
-2007-12,88
-2008-01,91.3
-2008-02,91.3
-2008-03,91.3
-2008-04,95.4
-2008-05,95.4
-2008-06,95.4
-2008-07,99.4
-2008-08,99.4
-2008-09,99.4
-2008-10,100.8
-2008-11,100.8
-2008-12,100.8
-2009-01,100
-2009-02,100
-2009-03,100
-2009-04,101.4
-2009-05,101.4
-2009-06,101.4
-2009-07,105
-2009-08,105
-2009-09,105
-2009-10,109
-2009-11,109
-2009-12,109
-2010-01,112.1
-2010-02,112.1
-2010-03,112.1
-2010-04,116.6
-2010-05,116.6
-2010-06,116.6
-2010-07,121.3
-2010-08,121.3
-2010-09,121.3
-2010-10,124.4
-2010-11,124.4
-2010-12,124.4
-2011-01,126.4
-2011-02,126.4
-2011-03,126.4
-2011-04,130.4
-2011-05,130.4
-2011-06,130.4
-2011-07,135.4
-2011-08,135.4
-2011-09,135.4
-2011-10,137.7
-2011-11,137.7
-2011-12,137.7
-2012-01,138.5
-2012-02,138.5
-2012-03,138.5
-2012-04,140.3
-2012-05,140.3
-2012-06,140.3
-2012-07,143.1
-2012-08,143.1
-2012-09,143.1
-2012-10,146.7
-2012-11,146.7
-2012-12,146.7
-2013-01,148.6
-2013-02,148.6
-2013-03,148.6
-2013-04,149.4
-2013-05,149.4
-2013-06,149.4
-2013-07,148.1
-2013-08,148.1
-2013-09,148.1
-2013-10,145.8
-2013-11,145.8
-2013-12,145.8
-2014-01,143.5
-2014-02,143.5
-2014-03,143.5
-2014-04,141.5
-2014-05,141.5
-2014-06,141.5
-2014-07,139.1
-2014-08,139.1
-2014-09,139.1
-2014-10,137
-2014-11,137
-2014-12,137
-2015-01,135.6
-2015-02,135.6
-2015-03,135.6
-2015-04,135
-2015-05,135
-2015-06,135
-2015-07,134.6
-2015-08,134.6
-2015-09,134.6
-2015-10,134.8
-2015-11,134.8
-2015-12,134.8
-2016-01,134.7
-2016-02,134.7
-2016-03,134.7
-2016-04,134.7
-2016-05,134.7
-2016-06,134.7
-2016-07,134.7
-2016-08,134.7
-2016-09,134.7
-2016-10,134.6
-2016-11,134.6
-2016-12,134.6
-2017-01,133.9
-2017-02,133.9
-2017-03,133.9
-2017-04,133.7
-2017-05,133.7
-2017-06,133.7
-2017-07,132.8
-2017-08,132.8
-2017-09,132.8
-2017-10,132.6
-2017-11,132.6
-2017-12,132.6
-2018-01,131.6
-2018-02,131.6
-2018-03,131.6
-2018-04,131.7
-2018-05,131.7
-2018-06,131.7
-2018-07,131.6
-2018-08,131.6
-2018-09,131.6
-2018-10,131.4
-2018-11,131.4
-2018-12,131.4
-2019-01,131
-2019-02,131
-2019-03,131
-2019-04,130.8
-2019-05,130.8
-2019-06,130.8
-2019-07,130.9
-2019-08,130.9
-2019-09,130.9
-2019-10,131.5
-2019-11,131.5
-2019-12,131.5
-2020-01,131.5
-2020-02,131.5
-2020-03,131.5
-2020-04,131.9
-2020-05,131.9
-2020-06,131.9
-2020-07,133.9
-2020-08,133.9
-2020-09,133.9
-2020-10,138.1
-2020-11,138.1
-2020-12,138.1
-2021-01,142.2
-2021-02,142.2
-2021-03,142.2
-2021-04,146.4
-2021-05,146.4
-2021-06,146.4
-2021-07,150.6
-2021-08,150.6
-2021-09,150.6
-2021-10,155.7
-2021-11,155.7
-2021-12,155.7
-2022-01,159.5
-2022-02,159.5
-2022-03,159.5
-2022-04,163.9
-2022-05,163.9
-2022-06,163.9
-2022-07,168.1
-2022-08,168.1
-2022-09,168.1
-2022-10,172.6
-2022-11,172.6
-2022-12,172.6
-2023-01,176.9
-2023-02,176.9
-2023-03,176.9
