--- Project_01/data/model_evaluation.csv
@@ -1,5 +1,0 @@
-,Train,Test
-Root Mean Squared Error,63733.52,63566.5
-Mean Absolute Error,46195.35,46109.85
-Mean Absolute Percentage Error,7.5,7.5
-R2 score,0.9,0.9
